1.修改手机sos 猫端不闪灯的问题 2.进入通话记录详情界面，右上角电量、入网等图标未刷新，显示红色 3.修改服务端代码 解决问题

+++ code/QzyTianTongServer/app/src/main/java/com/qzy/tiantong/service/utils/ModuleDormancyUtil.java
@@ -21,6 +21,10 @@ public class ModuleDormancyUtil {
     public static String getNodeString(String path) {
         String prop = "1";// 默认值
         try {
+            File file = new File(path);
+            if(!file.exists()){
+                return "-1";
+            }
             FileReader fileReader = new FileReader(path);
             BufferedReader reader = new BufferedReader(fileReader);
             prop = reader.readLine();
@@ -163,6 +167,9 @@ public class ModuleDormancyUtil {
                         OutputStreamWriter outputWrite = null;
                         PrintWriter print = null;
                         File file = new File(path);
+                        if(!file.exists()){
+                            break;
+                        }
                         Log.i("ModuleDormancyUtil", "write file[" + path + "] with value (" + state + ")");
 
                         output = new FileOutputStream(file);
